Create README.md
The BGU-MF test is a computerized math fluency test. The test includes simple math exercises of the four operations: addition, subtraction, multiplication, and division. Exercises in addition and subtraction used numbers between 0 and 20 (e.g., 7 + 6). Multiplication and division exercises were taken from the multiplication table (up to 10 X 10; e.g., 56÷ 7). The exercises appeared one by one on the screen, and participants asked to key in their answers. 
The presented exercise selected randomly from a pool of math facts (a total of 900 options: 440 addition exercises, 230 subtraction exercises, 120 multiplication exercises and 110 division exercises). 
In each trial, one exercise presented and participants were instructed to type their answers inside a square and press the enter key in order to proceed to the next trial. The typed answer presented on the screen.  The length of the test was 180 seconds. 
At the end of the test, a message of completion of the test appeared on the screen. The output for the performance includes a table with a line for each trial with the following details: the exercise that appeared, the participant’s response, accuracy (whether the participant’s response was correct or nor) and the response time per specific exercise. In addition, a summary of the total number of solved exercises accuracy (percent correct) per operation is presented, and average RT (for correct responses) per operation 
Experimenter may adjust the operation presented in the test, and whether if the test incudes the operand zero or not. 
Example of the test: yardengliksman.github.io/BGU-MF/
Instruction for participants: 
"Please answer as accurately and quickly as possible on the presented exercises."

Citations:

Gliksman, Y., Berebi, S., Hershman, R., & Henik, A. (2020). BGU-MF: Ben-Gurion University Math Fluency Test. Manuscript under review.